Fix #66 move validator to utils
