+++ rslaser/utils/validator.py
